+++ ChangeLog.md
@@ -1,5 +1,11 @@
 # OldNews ChangeLog
 
+## Unreleased
+
+**Released: WiP**
+
+- Tidy up the display and focus when the article list goes empty.
+
 ## v0.3.0
 
 **Released: 2026-01-19**